[movie-list-page]: move move all to movie-list-page
